2.2.6 ONLY USE WITH TT-RSS COMMIT d63329baa1 OR NEWER! - Update_tt-rss: Do not include the plugin names list when calculating article hashes   causes the option "Mark updated articles unread" to be triggered only when   the source has been updated, not when you change your plugin configuration   WARNING: THIS WILL RECALCULATE HASHES AND COULD MARK ARTICLES AS UNREAD - Update_tt-rss: Modify line_scroll_offset for scrolling with arrow up/down - Update_tt-rss: changed 21px to 19px margin to account for tt-rss scroll changes - Update_tt-rss: Check if find/replace changes were made and if changes were saved - Update_tt-rss: Add comment to changed lines

+++ update_tt-rss.php
@@ -1,22 +1,31 @@
 <?php
 /*	A simple script to update Tiny Tiny RSS (upload, extract and clean up).
  *	It takes a master.zip, i.e. a compressed snapshot from the master branch.
+ *	Tweaks:
+ *	- Adds <g>, <main> and <article> to the allowed elements in articles
+ *	- Adds a 19px margin for when scrolling to next/previous article
+ *  - Do not include the plugin names list when calculating article hashes:
+ *    causes the option "Mark updated articles unread" to be triggered only when
+ *    the source has been updated, not when you change your plugin configuration
+ *    WARNING: THIS WILL RECALCULATE HASHES AND COULD MARK ARTICLES AS UNREAD
+ *  - Modify line_scroll_offset for scrolling with arrow up/down
  *	Removal:
  *	- Removes useless folders: install, tests, feed-icons (moved to cache/feed-icons)
  *	- Removes useless files: .empty, .gitignore, *.less, *.map etc.
+ *	- Removes all language files, except for those set in $keep_langs / $keep_locale
  *	- Removes plugins, except for those set in $keep_plugins
- *	- Removes all language files, except for those set in $keep_languages / $keep_locale
  *	- Unlinks light.css.less mapping in light.css (prevents console error)
- *	Tweaks:
- *	- Adds <g>, <main> and <article> to the allowed elements in articles
- *	- Adds a 21px margin for when scrolling to next/previous article
  */
 
-$password = '';	// sha256 hash
-$keep_languages = ['en', 'nl'];
-$keep_locale = ['nl_NL'];
-$keep_plugins = ['af_readability', 'af_redditimgur', 'af_proxy_http', 'auth_internal', 'bookmarklets', 'note', 'share', 'vf_shared'];
-$root = pathinfo(__FILE__, PATHINFO_DIRNAME) . '/tt-rss';	// folder from extracted zip
+$password     = '';	// sha256 hash
+// Tweaks
+$no_plugins_hash = FALSE; // use FALSE to disable changes
+$line_offset     = 240;  // use FALSE to disable changes
+// Removal
+$keep_langs      = ['en', 'nl']; // use FALSE to disable
+$keep_locale     = ['nl_NL'];    // use FALSE to disable
+$keep_plugins    = ['af_readability', 'af_redditimgur', 'af_proxy_http', 'auth_internal', 'bookmarklets', 'note', 'share', 'vf_shared']; // use FALSE to disable
+$root            = pathinfo(__FILE__, PATHINFO_DIRNAME) . '/tt-rss'; // folder from extracted zip
 
 function remove($path, $key = null, $print = true) {
 	chdir($GLOBALS['root']);
@@ -45,11 +54,14 @@ function clean($dir = false, $keep, $ext = false) {
 function fart($file, $find, $replace) {
 	chdir($GLOBALS['root']);
 	$contents = file_get_contents($file);
-	$contents = str_replace($find, $replace, $contents);
-	file_put_contents($file, $contents);
+	$newcontents = str_replace($find, $replace, $contents);
+	if ($newcontents == $contents)
+		echo ' - <b>FAILED: No changes made.</b>';
+	else if (!file_put_contents($file, $newcontents))
+		echo ' - <b>FAILED: Could not save changes.</b>';
 }
 
-if(isset($_POST['submit'])) {
+if (isset($_POST['submit'])) {
 	echo '<style>ul{columns:3} ul>li{font-size:.9rem}</style>';
 	if (!empty($password) && (!isset($_POST['password']) || hash('sha256', $_POST['password']) != $password))
 		die('Password incorrect');
@@ -81,35 +93,49 @@ if(isset($_POST['submit'])) {
 		echo '<li>Unlinking .less source mapping in light.css</li>';
 		fart('themes/light.css', '/*# sourceMappingURL=light.css.map */', '');
 		echo '<li>Adding &lt;g&gt;, &lt;main&gt; and &lt;article&gt; to allowed elements for TorrentFreak and New Scientist articles</li>';
-		fart('include/functions.php', '$allowed_elements = array(', '$allowed_elements = array(\'g\', \'main\', \'article\', ');
+		fart('include/functions.php', '$allowed_elements = array(', '/* Changed by tt-rss updater script */ $allowed_elements = array(\'g\', \'main\', \'article\', ');
 		echo '<li>Adding margin for scrolling to articles</li>';
-		fart('js/Article.js', 'ctr.scrollTop = e.offsetTop;', 'ctr.scrollTop = e.offsetTop - (App.getInitParam("cdm_expanded") ? 21 : 0);');
+		fart('js/Article.js', 'ctr.scrollTop = row.offsetTop;', '/* Changed by tt-rss updater script */ ctr.scrollTop = row.offsetTop - (App.getInitParam("cdm_expanded") ? 19 : 0);');
+
+		if ($line_offset) {
+			echo '<li>Changing line offset for scrolling with cursor keys to '. $line_offset .'px</li>';
+			fart('js/Headlines.js', 'line_scroll_offset: 120', '/* Changed by tt-rss updater script */ line_scroll_offset: '. $line_offset);
+		} else echo '<li><b>Skipping</b> line offset change</li>';
+		
+		if ($no_plugins_hash) {
+			echo '<li>Changing the way article hashes are calculated: plugin names list is excluded.</li>';
+			fart('classes/rssutils.php', 'return sha1(implode(",", $pluginhost->get_plugin_names()) . $tmp)', '/* Changed by tt-rss updater script */ return sha1($tmp)');
+		} else echo '<li><b>Skipping</b> article hash change: plugin names list is still used to calculate hash.</li>';
 
 		echo '<li>Removing useless files...</li><ul>';
-		foreach(glob('{,*,*/*,*/*/*,*/*/*/*,*/*/*/*/*}/{.empty,.gitignore,*.less,*.map}', GLOB_BRACE) as $file)	// No spaces after comma between {}!
+		foreach(glob('{,*,*/*,*/*/*,*/*/*/*,*/*/*/*/*}/{.empty,.gitignore,*.less,*.map}', GLOB_BRACE) as $file) // No spaces after comma between {}!
 			remove($file);
 		remove('.editorconfig');
 		remove('.gitignore');
-//	remove('.gitlab-ci.yml');
 		remove('CONTRIBUTING.md');
 		remove('COPYING');
 		remove('README.md');
 		remove('feed-icons');
 		remove('install');
-//	remove('tests');
 		remove('utils');
 
-		echo '</ul><li>Removing unused plugins...</li><ul>';
-		clean('plugins', $keep_plugins);
-		
-		echo '</ul><li>Removing unused languages (all but '. implode(', ', $keep_locale) .', '. implode(', ', $keep_languages) .')...</li><ul>';
-		clean('locale', $keep_locale);
-		foreach(glob('{,*,*/*,*/*/*}/nls', GLOB_BRACE|GLOB_ONLYDIR) as $dir)
-			clean($dir, $keep_languages);
-		$keep_nls = ['colors.js', 'tt-rss-layer_ROOT.js', 'tt-rss-layer_en-us.js'];
-		foreach ($keep_locale as $l)
-			array_push($keep_nls, 'tt-rss-layer_'. str_replace('_', '-', strtolower($l)) .'.js');
-		clean('lib/dojo/nls', $keep_nls, '.js');
+		if (is_array($keep_langs)) {
+			echo '</ul><li>Removing unused languages (all but '. implode(', ', $keep_locale) .', '. implode(', ', $keep_langs) .')...</li><ul>';
+			clean('locale', $keep_locale);
+			foreach(glob('{,*,*/*,*/*/*}/nls', GLOB_BRACE|GLOB_ONLYDIR) as $dir)
+				clean($dir, $keep_langs);
+			$keep_nls = ['colors.js', 'tt-rss-layer_ROOT.js', 'tt-rss-layer_en-us.js'];
+		} else echo '<li><b>Skipping</b> language removal</li>';
+		if (is_array($keep_locale)) {
+			foreach ($keep_locale as $l)
+				array_push($keep_nls, 'tt-rss-layer_'. str_replace('_', '-', strtolower($l)) .'.js');
+			clean('lib/dojo/nls', $keep_nls, '.js');
+		} else echo '<li><b>Skipping</b> locale removal</li>';
+
+		if (is_array($keep_plugins)) {
+			echo '</ul><li>Removing unused plugins...</li><ul>';
+			clean('plugins', $keep_plugins);
+		} else echo '<li><b>Skipping</b> plugins removal</li>';
 
 		echo '</ul><li>Moving files into place...</li><ul>';
 		print_r(shell_exec('cp -Rf '. $GLOBALS['root'] .'/* '. pathinfo(__FILE__, PATHINFO_DIRNAME) .'/'));
